LayoutTests/canvas/philip/tests/2d.path.arcTo.transformation-expected.txt

< [index] >
2d.path.arcTo.transformation
arcTo joins up to the last subpath point correctly
References:
Actual output:
Expected output:

Passed
